Consolidate password and two-factor settings into a unified security page
The password update and two-factor authentication settings were separate
pages with their own routes and controllers. This merges them into a
single security settings page across all kit variants (Inertia
React/Vue/Svelte and Livewire), reducing navigation complexity and
aligning the settings layout.
