--- kits/Inertia/Fortify/Base/tests/Feature/Settings/TwoFactorAuthenticationTest.php
@@ -1,84 +1,0 @@
-<?php
-
-namespace Tests\Feature\Settings;
-
-use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Inertia\Testing\AssertableInertia as Assert;
-use Laravel\Fortify\Features;
-use Tests\TestCase;
-
-class TwoFactorAuthenticationTest extends TestCase
-{
-    use RefreshDatabase;
-
-    public function test_two_factor_settings_page_can_be_rendered()
-    {
-        $this->skipUnlessFortifyFeature(Features::twoFactorAuthentication());
-
-        Features::twoFactorAuthentication([
-            'confirm' => true,
-            'confirmPassword' => true,
-        ]);
-
-        $user = User::factory()->create();
-
-        $this->actingAs($user)
-            ->withSession(['auth.password_confirmed_at' => time()])
-            ->get(route('two-factor.show'))
-            ->assertInertia(fn (Assert $page) => $page
-                ->component('{{two_factor_settings}}')
-                ->where('twoFactorEnabled', false),
-            );
-    }
-
-    public function test_two_factor_settings_page_requires_password_confirmation_when_enabled()
-    {
-        $this->skipUnlessFortifyFeature(Features::twoFactorAuthentication());
-
-        $user = User::factory()->create();
-
-        Features::twoFactorAuthentication([
-            'confirm' => true,
-            'confirmPassword' => true,
-        ]);
-
-        $response = $this->actingAs($user)
-            ->get(route('two-factor.show'));
-
-        $response->assertRedirect(route('password.confirm'));
-    }
-
-    public function test_two_factor_settings_page_does_not_requires_password_confirmation_when_disabled()
-    {
-        $this->skipUnlessFortifyFeature(Features::twoFactorAuthentication());
-
-        $user = User::factory()->create();
-
-        Features::twoFactorAuthentication([
-            'confirm' => true,
-            'confirmPassword' => false,
-        ]);
-
-        $this->actingAs($user)
-            ->get(route('two-factor.show'))
-            ->assertOk()
-            ->assertInertia(fn (Assert $page) => $page
-                ->component('{{two_factor_settings}}'),
-            );
-    }
-
-    public function test_two_factor_settings_page_returns_forbidden_response_when_two_factor_is_disabled()
-    {
-        $this->skipUnlessFortifyFeature(Features::twoFactorAuthentication());
-
-        config(['fortify.features' => []]);
-
-        $user = User::factory()->create();
-
-        $this->actingAs($user)
-            ->withSession(['auth.password_confirmed_at' => time()])
-            ->get(route('two-factor.show'))
-            ->assertForbidden();
-    }
-}
