--- kits/Inertia/Fortify/Svelte/resources/js/pages/settings/Password.svelte
@@ -1,103 +1,0 @@
-<script lang="ts">
-    import { Form } from '@inertiajs/svelte';
-    import PasswordController from '@/actions/App/Http/Controllers/Settings/PasswordController';
-    import AppHead from '@/components/AppHead.svelte';
-    import Heading from '@/components/Heading.svelte';
-    import InputError from '@/components/InputError.svelte';
-    import PasswordInput from '@/components/PasswordInput.svelte';
-    import { Button } from '@/components/ui/button';
-    import { Label } from '@/components/ui/label';
-    import AppLayout from '@/layouts/AppLayout.svelte';
-    import SettingsLayout from '@/layouts/settings/Layout.svelte';
-    import { edit } from '@/routes/user-password';
-    import type { BreadcrumbItem } from '@/types';
-
-    const breadcrumbItems: BreadcrumbItem[] = [
-        {
-            title: 'Password settings',
-            href: edit(),
-        },
-    ];
-</script>
-
-<AppHead title="Password settings" />
-
-<AppLayout breadcrumbs={breadcrumbItems}>
-    <h1 class="sr-only">Password settings</h1>
-
-    <SettingsLayout>
-        <div class="space-y-6">
-            <Heading
-                variant="small"
-                title="Update password"
-                description="Ensure your account is using a long, random password to stay secure"
-            />
-
-            <Form
-                {...PasswordController.update.form()}
-                class="space-y-6"
-                options={{ preserveScroll: true }}
-                resetOnSuccess
-                resetOnError={[
-                    'password',
-                    'password_confirmation',
-                    'current_password',
-                ]}
-            >
-                {#snippet children({ errors, processing, recentlySuccessful })}
-                    <div class="grid gap-2">
-                        <Label for="current_password">Current password</Label>
-                        <PasswordInput
-                            id="current_password"
-                            name="current_password"
-                            class="mt-1 block w-full"
-                            autocomplete="current-password"
-                            placeholder="Current password"
-                        />
-                        <InputError message={errors.current_password} />
-                    </div>
-
-                    <div class="grid gap-2">
-                        <Label for="password">New password</Label>
-                        <PasswordInput
-                            id="password"
-                            name="password"
-                            class="mt-1 block w-full"
-                            autocomplete="new-password"
-                            placeholder="New password"
-                        />
-                        <InputError message={errors.password} />
-                    </div>
-
-                    <div class="grid gap-2">
-                        <Label for="password_confirmation"
-                            >Confirm password</Label
-                        >
-                        <PasswordInput
-                            id="password_confirmation"
-                            name="password_confirmation"
-                            class="mt-1 block w-full"
-                            autocomplete="new-password"
-                            placeholder="Confirm password"
-                        />
-                        <InputError message={errors.password_confirmation} />
-                    </div>
-
-                    <div class="flex items-center gap-4">
-                        <Button
-                            type="submit"
-                            disabled={processing}
-                            data-test="update-password-button"
-                        >
-                            Save password
-                        </Button>
-
-                        {#if recentlySuccessful}
-                            <p class="text-sm text-neutral-600">Saved.</p>
-                        {/if}
-                    </div>
-                {/snippet}
-            </Form>
-        </div>
-    </SettingsLayout>
-</AppLayout>
