+++ kits/Inertia/Fortify/Svelte/resources/js/pages/settings/Security.svelte
@@ -3,16 +3,21 @@
     import ShieldBan from 'lucide-svelte/icons/shield-ban';
     import ShieldCheck from 'lucide-svelte/icons/shield-check';
     import { onDestroy } from 'svelte';
+    import SecurityController from '@/actions/App/Http/Controllers/Settings/SecurityController';
     import AppHead from '@/components/AppHead.svelte';
     import Heading from '@/components/Heading.svelte';
+    import InputError from '@/components/InputError.svelte';
+    import PasswordInput from '@/components/PasswordInput.svelte';
     import TwoFactorRecoveryCodes from '@/components/TwoFactorRecoveryCodes.svelte';
     import TwoFactorSetupModal from '@/components/TwoFactorSetupModal.svelte';
     import { Badge } from '@/components/ui/badge';
     import { Button } from '@/components/ui/button';
+    import { Label } from '@/components/ui/label';
     import AppLayout from '@/layouts/AppLayout.svelte';
     import SettingsLayout from '@/layouts/settings/Layout.svelte';
     import { twoFactorAuthState } from '@/lib/twoFactorAuth.svelte';
-    import { disable, enable, show } from '@/routes/two-factor';
+    import { edit } from '@/routes/security';
+    import { disable, enable } from '@/routes/two-factor';
     import type { BreadcrumbItem } from '@/types';
 
     let {
@@ -25,25 +30,97 @@
 
     const breadcrumbs: BreadcrumbItem[] = [
         {
-            title: 'Two-factor authentication',
-            href: show(),
+            title: 'Security settings',
+            href: edit(),
         },
     ];
 
     const twoFactorAuth = twoFactorAuthState();
     let showSetupModal = $state(false);
 
-    onDestroy(() => {
-        twoFactorAuth.clearTwoFactorAuthData();
-    });
+    onDestroy(() => twoFactorAuth.clearTwoFactorAuthData());
 </script>
 
-<AppHead title="Two-factor authentication" />
+<AppHead title="Security settings" />
 
 <AppLayout {breadcrumbs}>
-    <h1 class="sr-only">Two-factor authentication settings</h1>
+    <h1 class="sr-only">Security settings</h1>
 
     <SettingsLayout>
+        <div class="space-y-6">
+            <Heading
+                variant="small"
+                title="Update password"
+                description="Ensure your account is using a long, random password to stay secure"
+            />
+
+            <Form
+                {...SecurityController.update.form()}
+                class="space-y-6"
+                options={{ preserveScroll: true }}
+                resetOnSuccess
+                resetOnError={[
+                    'password',
+                    'password_confirmation',
+                    'current_password',
+                ]}
+            >
+                {#snippet children({ errors, processing, recentlySuccessful })}
+                    <div class="grid gap-2">
+                        <Label for="current_password">Current password</Label>
+                        <PasswordInput
+                            id="current_password"
+                            name="current_password"
+                            class="mt-1 block w-full"
+                            autocomplete="current-password"
+                            placeholder="Current password"
+                        />
+                        <InputError message={errors.current_password} />
+                    </div>
+
+                    <div class="grid gap-2">
+                        <Label for="password">New password</Label>
+                        <PasswordInput
+                            id="password"
+                            name="password"
+                            class="mt-1 block w-full"
+                            autocomplete="new-password"
+                            placeholder="New password"
+                        />
+                        <InputError message={errors.password} />
+                    </div>
+
+                    <div class="grid gap-2">
+                        <Label for="password_confirmation"
+                            >Confirm password</Label
+                        >
+                        <PasswordInput
+                            id="password_confirmation"
+                            name="password_confirmation"
+                            class="mt-1 block w-full"
+                            autocomplete="new-password"
+                            placeholder="Confirm password"
+                        />
+                        <InputError message={errors.password_confirmation} />
+                    </div>
+
+                    <div class="flex items-center gap-4">
+                        <Button
+                            type="submit"
+                            disabled={processing}
+                            data-test="update-password-button"
+                        >
+                            Save password
+                        </Button>
+
+                        {#if recentlySuccessful}
+                            <p class="text-sm text-neutral-600">Saved.</p>
+                        {/if}
+                    </div>
+                {/snippet}
+            </Form>
+        </div>
+
         <div class="space-y-6">
             <Heading
                 variant="small"
