+++ kits/Inertia/Fortify/Vue/resources/js/pages/settings/Security.vue
@@ -2,15 +2,20 @@
 import { Form, Head } from '@inertiajs/vue3';
 import { ShieldBan, ShieldCheck } from 'lucide-vue-next';
 import { onUnmounted, ref } from 'vue';
+import SecurityController from '@/actions/App/Http/Controllers/Settings/SecurityController';
 import Heading from '@/components/Heading.vue';
+import InputError from '@/components/InputError.vue';
+import PasswordInput from '@/components/PasswordInput.vue';
 import TwoFactorRecoveryCodes from '@/components/TwoFactorRecoveryCodes.vue';
 import TwoFactorSetupModal from '@/components/TwoFactorSetupModal.vue';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
+import { Label } from '@/components/ui/label';
 import { useTwoFactorAuth } from '@/composables/useTwoFactorAuth';
 import AppLayout from '@/layouts/AppLayout.vue';
 import SettingsLayout from '@/layouts/settings/Layout.vue';
-import { disable, enable, show } from '@/routes/two-factor';
+import { edit } from '@/routes/security';
+import { disable, enable } from '@/routes/two-factor';
 import type { BreadcrumbItem } from '@/types';
 
 type Props = {
@@ -25,26 +30,107 @@ withDefaults(defineProps<Props>(), {
 
 const breadcrumbs: BreadcrumbItem[] = [
     {
-        title: 'Two-factor authentication',
-        href: show(),
+        title: 'Security settings',
+        href: edit(),
     },
 ];
 
 const { hasSetupData, clearTwoFactorAuthData } = useTwoFactorAuth();
 const showSetupModal = ref<boolean>(false);
 
-onUnmounted(() => {
-    clearTwoFactorAuthData();
-});
+onUnmounted(() => clearTwoFactorAuthData());
 </script>
 
 <template>
     <AppLayout :breadcrumbs="breadcrumbs">
-        <Head title="Two-factor authentication" />
+        <Head title="Security settings" />
 
-        <h1 class="sr-only">Two-factor authentication settings</h1>
+        <h1 class="sr-only">Security settings</h1>
 
         <SettingsLayout>
+            <div class="space-y-6">
+                <Heading
+                    variant="small"
+                    title="Update password"
+                    description="Ensure your account is using a long, random password to stay secure"
+                />
+
+                <Form
+                    v-bind="SecurityController.update.form()"
+                    :options="{
+                        preserveScroll: true,
+                    }"
+                    reset-on-success
+                    :reset-on-error="[
+                        'password',
+                        'password_confirmation',
+                        'current_password',
+                    ]"
+                    class="space-y-6"
+                    v-slot="{ errors, processing, recentlySuccessful }"
+                >
+                    <div class="grid gap-2">
+                        <Label for="current_password">Current password</Label>
+                        <PasswordInput
+                            id="current_password"
+                            name="current_password"
+                            class="mt-1 block w-full"
+                            autocomplete="current-password"
+                            placeholder="Current password"
+                        />
+                        <InputError :message="errors.current_password" />
+                    </div>
+
+                    <div class="grid gap-2">
+                        <Label for="password">New password</Label>
+                        <PasswordInput
+                            id="password"
+                            name="password"
+                            class="mt-1 block w-full"
+                            autocomplete="new-password"
+                            placeholder="New password"
+                        />
+                        <InputError :message="errors.password" />
+                    </div>
+
+                    <div class="grid gap-2">
+                        <Label for="password_confirmation"
+                            >Confirm password</Label
+                        >
+                        <PasswordInput
+                            id="password_confirmation"
+                            name="password_confirmation"
+                            class="mt-1 block w-full"
+                            autocomplete="new-password"
+                            placeholder="Confirm password"
+                        />
+                        <InputError :message="errors.password_confirmation" />
+                    </div>
+
+                    <div class="flex items-center gap-4">
+                        <Button
+                            :disabled="processing"
+                            data-test="update-password-button"
+                            >Save password</Button
+                        >
+
+                        <Transition
+                            enter-active-class="transition ease-in-out"
+                            enter-from-class="opacity-0"
+                            leave-active-class="transition ease-in-out"
+                            leave-to-class="opacity-0"
+                        >
+                            <p
+                                v-show="recentlySuccessful"
+                                class="text-sm text-neutral-600"
+                            >
+                                Saved.
+                            </p>
+                        </Transition>
+                    </div>
+                </Form>
+            </div>
+
             <div class="space-y-6">
                 <Heading
                     variant="small"
