+++ kits/Inertia/Svelte/resources/js/layouts/settings/Layout.svelte
@@ -8,8 +8,7 @@
     import { toUrl } from '@/lib/utils';
     import { edit as editAppearance } from '@/routes/appearance';
     import { edit as editProfile } from '@/routes/profile';
-    import { show } from '@/routes/two-factor';
-    import { edit as editPassword } from '@/routes/user-password';
+    import { edit as editSecurity } from '@/routes/security';
     import type { NavItem } from '@/types';
 
     let {
@@ -24,12 +23,8 @@
             href: editProfile(),
         },
         {
-            title: 'Password',
-            href: editPassword(),
-        },
-        {
-            title: 'Two-factor auth',
-            href: show(),
+            title: 'Security',
+            href: editSecurity(),
         },
         {
             title: 'Appearance',
