+++ kits/Inertia/Vue/resources/js/layouts/settings/Layout.vue
@@ -7,8 +7,7 @@ import { useCurrentUrl } from '@/composables/useCurrentUrl';
 import { toUrl } from '@/lib/utils';
 import { edit as editAppearance } from '@/routes/appearance';
 import { edit as editProfile } from '@/routes/profile';
-import { show } from '@/routes/two-factor';
-import { edit as editPassword } from '@/routes/user-password';
+import { edit as editSecurity } from '@/routes/security';
 import type { NavItem } from '@/types';
 
 const sidebarNavItems: NavItem[] = [
@@ -17,12 +16,8 @@ const sidebarNavItems: NavItem[] = [
         href: editProfile(),
     },
     {
-        title: 'Password',
-        href: editPassword(),
-    },
-    {
-        title: 'Two-factor auth',
-        href: show(),
+        title: 'Security',
+        href: editSecurity(),
     },
     {
         title: 'Appearance',
