+++ kits/Livewire/Components/routes/settings.php
@@ -1,9 +1,8 @@
 <?php
 
 use App\Livewire\Settings\Appearance;
-use App\Livewire\Settings\Password;
 use App\Livewire\Settings\Profile;
-use App\Livewire\Settings\TwoFactor;
+use App\Livewire\Settings\Security;
 use Illuminate\Support\Facades\Route;
 use Laravel\Fortify\Features;
 
@@ -14,10 +13,9 @@ Route::middleware(['auth'])->group(function () {
 });
 
 Route::middleware(['auth', 'verified'])->group(function () {
-    Route::livewire('settings/password', Password::class)->name('user-password.edit');
     Route::livewire('settings/appearance', Appearance::class)->name('appearance.edit');
 
-    Route::livewire('settings/two-factor', TwoFactor::class)
+    Route::livewire('settings/security', Security::class)
         ->middleware(
             when(
                 Features::canManageTwoFactorAuthentication()
@@ -26,5 +24,5 @@ Route::middleware(['auth', 'verified'])->group(function () {
                 [],
             ),
         )
-        ->name('two-factor.show');
+        ->name('security.edit');
 });
